+++ Examples/Fading Demo/iCarouselExampleViewController.m
@@ -1,0 +1,137 @@
+//
+//  iCarouselExampleViewController.m
+//  iCarouselExample
+//
+//  Created by Nick Lockwood on 03/04/2011.
+//  Copyright 2011 Charcoal Design. All rights reserved.
+//
+
+#import "iCarouselExampleViewController.h"
+
+
+@interface iCarouselExampleViewController ()
+
+@property (nonatomic, retain) NSMutableArray *items;
+
+@end
+
+
+@implementation iCarouselExampleViewController
+
+@synthesize carousel;
+@synthesize items;
+
+- (void)awakeFromNib
+{
+    //set up data
+    //your carousel should always be driven by an array of
+    //data of some kind - don't store data in your item views
+    //or the recycling mechanism will destroy your data once
+    //your item views move off-screen
+    self.items = [NSMutableArray array];
+    for (int i = 0; i < 100; i++)
+    {
+        [items addObject:[NSNumber numberWithInt:i]];
+    }
+}
+
+- (void)dealloc
+{
+    //it's a good idea to set these to nil here to avoid
+    //sending messages to a deallocated viewcontroller
+    //this is true even if your project is using ARC, unless
+    //you are targeting iOS 5 as a minimum deployment target
+    carousel.delegate = nil;
+    carousel.dataSource = nil;
+    
+    [carousel release];
+    [items release];
+    [super dealloc];
+}
+
+#pragma mark -
+#pragma mark View lifecycle
+
+- (void)viewDidLoad
+{
+    [super viewDidLoad];
+    
+    //configure carousel
+    carousel.type = iCarouselTypeRotary;
+}
+
+- (void)viewDidUnload
+{
+    [super viewDidUnload];
+    
+    //free up memory by releasing subviews
+    self.carousel = nil;
+}
+
+- (BOOL)shouldAutorotateToInterfaceOrientation:(UIInterfaceOrientation)interfaceOrientation
+{
+    return YES;
+}
+
+#pragma mark -
+#pragma mark iCarousel methods
+
+- (NSUInteger)numberOfItemsInCarousel:(iCarousel *)carousel
+{
+    //return the total number of items in the carousel
+    return [items count];
+}
+
+- (UIView *)carousel:(iCarousel *)carousel viewForItemAtIndex:(NSUInteger)index reusingView:(UIView *)view
+{
+    UILabel *label = nil;
+    
+    //create new view if no view is available for recycling
+    if (view == nil)
+    {
+        //don't do anything specific to the index within
+        //this `if (view == nil) {...}` statement because the view will be
+        //recycled and used with other index values later
+        view = [[[UIImageView alloc] initWithFrame:CGRectMake(0, 0, 200.0f, 200.0f)] autorelease];
+        ((UIImageView *)view).image = [UIImage imageNamed:@"page.png"];
+        view.contentMode = UIViewContentModeCenter;
+        
+        label = [[[UILabel alloc] initWithFrame:view.bounds] autorelease];
+        label.backgroundColor = [UIColor clearColor];
+        label.textAlignment = UITextAlignmentCenter;
+        label.font = [label.font fontWithSize:50];
+        label.tag = 1;
+        [view addSubview:label];
+    }
+    else
+    {
+        //get a reference to the label in the recycled view
+        label = (UILabel *)[view viewWithTag:1];
+    }
+    
+    //set item label
+    //remember to always set any properties of your carousel item
+    //views outside of the `if (view == nil) {...}` check otherwise
+    //you'll get weird issues with carousel item content appearing
+    //in the wrong place in the carousel
+    label.text = [[items objectAtIndex:index] stringValue];
+    
+    return view;
+}
+
+- (CGFloat)carousel:(iCarousel *)carousel valueForOption:(iCarouselOption)option withDefault:(CGFloat)value
+{
+    switch (option)
+    {
+        case iCarouselOptionFadeMin:
+            return -0.2;
+        case iCarouselOptionFadeMax:
+            return 0.2;
+        case iCarouselOptionFadeRange:
+            return 2.0;
+        default:
+            return value;
+    }
+}
+
+@end